# Speed up gyro calibration