--- a/src/modules/sensors/sensor_params_gyro1.c
+++ b/src/modules/sensors/sensor_params_gyro1.c
@@ -72,26 +72,3 @@ PARAM_DEFINE_FLOAT(CAL_GYRO1_YOFF, 0.0f);
  */
 PARAM_DEFINE_FLOAT(CAL_GYRO1_ZOFF, 0.0f);
 
-/**
- * Gyro X-axis scaling factor
- *
- * @category system
- * @group Sensor Calibration
- */
-PARAM_DEFINE_FLOAT(CAL_GYRO1_XSCALE, 1.0f);
-
-/**
- * Gyro Y-axis scaling factor
- *
- * @category system
- * @group Sensor Calibration
- */
-PARAM_DEFINE_FLOAT(CAL_GYRO1_YSCALE, 1.0f);
-
-/**
- * Gyro Z-axis scaling factor
- *
- * @category system
- * @group Sensor Calibration
- */
-PARAM_DEFINE_FLOAT(CAL_GYRO1_ZSCALE, 1.0f);