… posix-configs/excelsior/px4.config