- speedup gyro calibration from 20s to 1s
- slightly stricter motion detection check (change the threshold from 0.573 deg to 0.4), as the existing one allowed for quite a bit of motion.
- remove unused gyro scale calibration parameters

The existing calibration collects 5000 samples, which is unnecessary, so this PR reduces it to 250 (1 second). Tests showed that the standard deviation of the estimated offsets over 10 runs is in the same range for 1s as for 20s (the increase in temperature has a bigger effect on the offset). This was tested with an MPU9250, ICM20608, ICM20689 and BMI088.

In case of motion, calibration is restarted automatically, maximally 30 times (same as existing behavior).